fix for int parsing

+++ backend/gunicorn_config.py
@@ -8,8 +8,8 @@ backlog = 2048
 worker_class = "gthread"
 timeout = 30
 keepalive = 2
-workers = multiprocessing.cpu_count() / 2
-threads = multiprocessing.cpu_count()
+workers = int(multiprocessing.cpu_count() / 2)
+threads = int(multiprocessing.cpu_count())
 
 proc_name = None
 daemon = False
